+++ src/columns.h
@@ -76,18 +76,19 @@ inline List create_columns(
 
   auto locale_info = std::make_shared<LocaleInfo>(locale);
 
-  std::vector<std::string> res_nms;
-
   size_t i = 0;
 
   bool add_filename = !Rf_isNull(id);
 
   List res(num_cols + add_filename);
 
+  CharacterVector res_nms(num_cols + add_filename);
+
   if (add_filename) {
-    res[i++] =
+    res[i] =
         generate_filename_column(filenames, idx->row_sizes(), idx->num_rows());
-    res_nms.push_back(Rcpp::as<std::string>(id));
+    res_nms[i] = Rcpp::as<Rcpp::CharacterVector>(id)[0];
+    ++i;
   }
 
   auto my_collectors = resolve_collectors(
@@ -124,7 +125,7 @@ inline List create_columns(
                                    locale_info,
                                    std::string()};
 
-    res_nms.push_back(collector.name());
+    res_nms[i] = collector.name();
 
     switch (collector.type()) {
     case column_type::Dbl:
@@ -232,6 +233,9 @@ inline List create_columns(
     // Resize the list appropriately
     SETLENGTH(res, i);
     SET_TRUELENGTH(res, i);
+
+    SETLENGTH(res_nms, i);
+    SET_TRUELENGTH(res_nms, i);
   }
 
   res.attr("names") = res_nms;